notebooks/requirements.txt
# Core Strands Agents
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Additional utilities (if needed by tools)
python-dotenv>=1.0.0 